+++ src/main/java/com/nali/list/netmethods/clientmessage/SetSEMap.java
@@ -14,7 +14,7 @@ import java.util.UUID;
 
 public class SetSEMap
 {
-    public static int ID = 0;
+    public static byte ID = 0;
 
     public static void run(ClientMessage clientmessage)
     {
