+++ src/main/java/com/nali/list/netmethods/clientmessage/SetTroublemaker.java
@@ -9,7 +9,7 @@ import static com.nali.small.gui.features.messages.inventory.TroublemakerGUIFeat
 
 public class SetTroublemaker
 {
-    public static int ID = 4;
+    public static byte ID = 4;
 
     public static void run(ClientMessage clientmessage)
     {
