+++ src/main/java/com/nali/list/netmethods/servermessage/SyncUUIDToClient.java
@@ -11,7 +11,7 @@ import net.minecraft.entity.player.EntityPlayerMP;
 
 public class SyncUUIDToClient
 {
-    public static int ID = 0;
+    public static byte ID = 0;
 
     public static void run(EntityPlayerMP entityplayermp, ServerMessage servermessage)
     {
